Add RPi Ansible configuration
We add Ansible playbooks that take care of initial configuration.

install_drivers.yaml takes care of LCD display driver installation.

init.yaml clones this repository to RPi and configures RPi so that the
RPi user is automatically logged in on startup and so that the publisher
starts running on startup.